src/Part1/requirements.txt
Flask==2.1.2
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
psycopg2-binary>=2.8.6
Werkzeug==2.0.3